# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kira Code API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/approvals/{id}/respond": {
      "post": {
        "tags": [
          "approvals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalResponse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse58"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalResponse": {
        "type": "object",
        "properties": {
          "execution_process_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalOutcome"
          }
        },
        "required": [
          "execution_process_id",
          "status"
        ]
      },
      "ApprovalOutcome": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "approved"
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "reason": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "status": {
                "type": "string",
                "const": "denied"
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "answers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/QuestionAnswer"
                }
              },
              "status": {
                "type": "string",
                "const": "answered"
              }
            },
            "required": [
              "status",
              "answers"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "timed_out"
              }
            },
            "required": [
              "status"
            ]
          }
        ]
      },
      "QuestionAnswer": {
        "description": "A question–answer pair. `answer` holds one or more selected labels/values.",
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "question",
          "answer"
        ]
      },
      "ApiResponse58": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}